Added -Wshorten-64-to-32 and fixed all warnings

+++ include/modules/dns.h
@@ -192,7 +192,7 @@ namespace DNS
 		/* Creator of this request */
 		Module* const creator;
 
-		Request(Manager* mgr, Module* mod, const std::string& addr, QueryType qt, bool usecache = true, unsigned int timeout = 0)
+		Request(Manager* mgr, Module* mod, const std::string& addr, QueryType qt, bool usecache = true, unsigned long timeout = 0)
 			: Timer(timeout ? timeout : ServerInstance->Config->ConfValue("dns")->getDuration("timeout", 5, 1))
 			, manager(mgr)
 			, question(addr, qt)
