+++ src/modules/m_callerid.cpp
@@ -178,7 +178,7 @@ class CommandAccept : public Command
 
 public:
 	CallerIDExtInfo extInfo;
-	unsigned int maxaccepts;
+	unsigned long maxaccepts;
 	CommandAccept(Module* Creator) : Command(Creator, "ACCEPT", 1),
 		extInfo(Creator)
 	{
@@ -279,7 +279,7 @@ public:
 		callerid_data* dat = extInfo.Get(user, true);
 		if (dat->accepting.size() >= maxaccepts)
 		{
-			user->WriteNumeric(ERR_ACCEPTFULL, InspIRCd::Format("Accept list is full (limit is %d)", maxaccepts));
+			user->WriteNumeric(ERR_ACCEPTFULL, InspIRCd::Format("Accept list is full (limit is %lu)", maxaccepts));
 			return false;
 		}
 		if (!dat->accepting.insert(whotoadd).second)
@@ -361,7 +361,7 @@ class ModuleCallerID
 
 	// Configuration variables:
 	bool tracknick; // Allow ACCEPT entries to update with nick changes.
-	unsigned int notify_cooldown; // Seconds between notifications.
+	unsigned long notify_cooldown; // Seconds between notifications.
 
 	/** Removes a user from all accept lists
 	 * @param who The user to remove from accepts
@@ -419,7 +419,7 @@ public:
 			time_t now = ServerInstance->Time();
 			/* +g and *not* accepted */
 			user->WriteNumeric(ERR_TARGUMODEG, dest->nick, "is in +g mode (server-side ignore).");
-			if (now > (dat->lastnotify + notify_cooldown))
+			if (now > (dat->lastnotify + long(notify_cooldown)))
 			{
 				user->WriteNumeric(RPL_TARGNOTIFY, dest->nick, "has been informed that you messaged them.");
 				dest->WriteRemoteNumeric(RPL_UMODEGMSG, user->nick, InspIRCd::Format("%s@%s", user->ident.c_str(), user->GetDisplayedHost().c_str()), InspIRCd::Format("is messaging you, and you have user mode +g set. Use /ACCEPT +%s to allow.",
