+++ src/modules/m_joinflood.cpp
@@ -155,7 +155,7 @@ class ModuleJoinFlood
 	void ReadConfig(ConfigStatus&) override
 	{
 		auto tag = ServerInstance->Config->ConfValue("joinflood");
-		duration = tag->getDuration("duration", 60, 10, 600);
+		duration = static_cast<unsigned int>(tag->getDuration("duration", 60, 10, 600));
 		bootwait = tag->getDuration("bootwait", 30);
 		splitwait = tag->getDuration("splitwait", 30);
